Changes to iterative robot generation.  Can now iteratively build a set of robots and merge the best robots into a final population.
I found a number of things that (I suspect) were not working as
intended.  One big issue discovered is that most populations quickly
became the size of the protected robots setting (20 by default given the
.02 setting and 1k population size).  This is because the robot killer
was always killing a portion of good robots by weight or age even if the
population was not full.  This meant every iteration non predicting
robots were killed then the remaining potentially useful set was culled
by weight/age. Ive changed this to only cull by weight and age when the
population is full, this allows more time for those robots to mutate and
grow and the effect is quite noticeable (my population are now much
bigger than just the 20 each round).  Also I randomly choose to cull by
age then weight or weight then age to give equal weight to these
options.

I also changed killNonPredicting robots to an integer.  0 means off.
Higher values indicate the robot must have made a prediction in the last
X time points.  So a setting of 90 on SPX would mean kill the robot if
it has not made a prediction for 90 days.  This seems to much better
cull robots that become inactive over time I defaulted this value to 120
in the sample config.

I also added a new setting ageBeforeKillingNonPredictingRobots  This is
the count of time points before a robot can be killed for not
predicting.  I found many cases were allowing a robot to run two or
three days made it start predicting.  I defaulted this value to 3 in the
sample config.

+++ src/main/java/com/alphatica/genotick/population/PopulationDAOFileSystem.java
@@ -72,7 +72,9 @@ public class PopulationDAOFileSystem implements PopulationDAO {
 
     @Override
     public void removeRobot(RobotName robotName) {
-        names.remove(robotName);
+        if(!names.remove(robotName)) {
+            throw new DAOException("Unable to remove robot " + robotName.toString());
+        }
         deleteFile(robotName);
     }
 
@@ -85,7 +87,7 @@ public class PopulationDAOFileSystem implements PopulationDAO {
     private void deleteFile(RobotName robotName) {
         File file = createFileForName(robotName);
         boolean result = file.delete();
-        if(!result)
+        if(!result && file.exists())
             throw new DAOException("Unable to remove file " + file.getAbsolutePath());
     }
